feature(Texture): 32 to 16 bits
Should reduce VRAM usage by 50%. We could save even more space with R11G11B10 format, eventually.

+++ src/Lumiere/GPU/Texture.cpp
@@ -293,13 +293,13 @@ namespace lum::gpu
             case DepthStencil:
                 return GL_DEPTH24_STENCIL8;
             case Red:
-                return type == GLUtils::Float ? GL_R32F : GL_R8;
+                return type == GLUtils::Float ? GL_R16F : GL_R8;
             case RG:
-                return type == GLUtils::Float ? GL_RG32F : GL_RG8;
+                return type == GLUtils::Float ? GL_RG16F : GL_RG8;
             case RGB:
-                return type == GLUtils::Float ? GL_RGB32F : GL_RGB8;
+                return type == GLUtils::Float ? GL_RGB16F : GL_RGB8;
             case RGBA:
-                return type == GLUtils::Float ? GL_RGBA32F : GL_RGBA8;
+                return type == GLUtils::Float ? GL_RGBA16F : GL_RGBA8;
             case SRGB:
                 return GL_SRGB8;
             default:
